requirements.txt
flask
gunicorn
python-dotenv
seleniumbase==4.37.1
PyAutoGUI
selenium
pytz
Werkzeug
click
itsdangerous
Jinja2
MarkupSafe
azure-data-tables
azure-storage-blob